refactor: conditional compile when glibc static-crt
